+++ src/ace_time_c/time_zone.h
@@ -7,7 +7,7 @@
  * @file time_zone.h
  *
  * A class representing a specific IANA time zone, containing a pointer to a
- * ZoneInfo data containing the TZ database info, and a ZoneProcessing instance
+ * ZoneInfo data containing the TZ database info, and a ZoneProcessor instance
  * which calculates DST offsets from the ZoneInfo.
  */
 
@@ -15,7 +15,7 @@
 #define ACE_TIME_C_TIME_ZONE_H
 
 #include "zone_info.h"
-#include "zone_processing.h"
+#include "zone_processor.h"
 
 #ifdef __cplusplus
 extern "C" {
@@ -23,7 +23,7 @@ extern "C" {
 
 typedef struct AtcTimeZone {
   const AtcZoneInfo *zone_info;
-  AtcZoneProcessing *zone_processing;
+  AtcZoneProcessor *zone_processor;
 } AtcTimeZone;
 
 #ifdef __cplusplus